refactor: move utils to its own folder
