+++ PGLib/include/util/NIFUtil.hpp
