+++ PGLib/src/ModManagerDirectory.cpp
@@ -14,7 +14,7 @@
 
 #include "BethesdaDirectory.hpp"
 #include "PGGlobals.hpp"
-#include "ParallaxGenUtil.hpp"
+#include "util/ParallaxGenUtil.hpp"
 
 using namespace std;
