+++ PGLib/src/patchers/PatcherMeshShaderVanillaParallax.cpp
@@ -3,8 +3,8 @@
 #include <Geometry.hpp>
 #include <boost/algorithm/string.hpp>
 
-#include "Logger.hpp"
-#include "NIFUtil.hpp"
+#include "util/Logger.hpp"
+#include "util/NIFUtil.hpp"
 
 using namespace std;
